Yassine branch

+++ task-service/src/controllers/sectionController.ts
@@ -2,19 +2,16 @@ import { Request, Response } from "express";
 import asyncHandler from "express-async-handler";
 import Section from "../models/Section";
 import { ISection } from "../interfaces/Section";
+import mongoose from "mongoose";
 
 // @desc    Get room sections
 // @route   POST /api/v1/sections/
 // @access  Private
 const getRoomSections = asyncHandler(async (req: Request, res: Response) => {
   const { roomId } = req.query;
-  try {
-    const sections = await Section.find({ roomId });
+  const sections = await Section.find({ roomId });
 
-    res.json(sections);
-  } catch (e) {
-    res.status(404).send({ error: "room not found" });
-  }
+  res.json(sections);
 });
 
 // @desc    Get section
@@ -22,14 +19,16 @@ const getRoomSections = asyncHandler(async (req: Request, res: Response) => {
 // @access  Private
 const getSection = asyncHandler(async (req: Request, res: Response) => {
   const { sectionId } = req.params;
-
-  try {
-    const section = await Section.findById(sectionId).populate("tasks");
-    if (!section) res.status(404).send({ error: "section not found" });
-    res.json(section);
-  } catch (e) {
-    res.status(400).send({ error: "id format not correct" });
+  if (!mongoose.Types.ObjectId.isValid(sectionId)) {
+    res.status(400);
+    throw Error("Id format not correct");
   }
+  const section = await Section.findById(sectionId).populate("tasks");
+  if (!section) {
+    res.status(404);
+    throw Error("section not found");
+  }
+  res.json(section);
 });
 
 // @desc    Create a section
@@ -37,16 +36,16 @@ const getSection = asyncHandler(async (req: Request, res: Response) => {
 // @access  Private
 const createSection = asyncHandler(async (req: Request, res: Response) => {
   const { roomId, name } = req.body;
-  try {
-    const section: ISection = await Section.create({
-      roomId,
-      name,
-    });
-
-    res.status(201).json(section);
-  } catch {
-    res.status(400).send({ error: "roomId and name is required" });
+  if (!roomId || !name) {
+    res.status(400);
+    throw Error("roomId and name is required");
   }
+  const section: ISection = await Section.create({
+    roomId,
+    name,
+  });
+
+  res.status(201).json(section);
 });
 
 // @desc    Update a section
@@ -55,19 +54,24 @@ const createSection = asyncHandler(async (req: Request, res: Response) => {
 const updateSection = asyncHandler(async (req: Request, res: Response) => {
   const { name } = req.body;
   const { sectionId } = req.params;
+  if (!mongoose.Types.ObjectId.isValid(sectionId)) {
+    res.status(400);
+    throw Error("Id format not correct");
+  }
+  if (!name) {
+    res.status(400);
+    throw Error("name is required");
+  }
 
-  try {
-    const section = await Section.findById(sectionId);
-    if (section) {
-      section.name = name;
+  const section = await Section.findById(sectionId);
+  if (section) {
+    section.name = name;
 
-      const updatedSection = await section.save();
-      res.json(updatedSection);
-    } else {
-      res.status(404).send({ error: "section not found" });
-    }
-  } catch {
-    res.status(400).send({ error: "roomId and name is required" });
+    const updatedSection = await section.save();
+    res.json(updatedSection);
+  } else {
+    res.status(404);
+    throw Error("section not found");
   }
 });
 
@@ -77,14 +81,19 @@ const updateSection = asyncHandler(async (req: Request, res: Response) => {
 const deleteSection = asyncHandler(async (req: Request, res: Response) => {
   const { sectionId } = req.params;
 
-  try {
-    await Section.findByIdAndDelete(sectionId);
-
-    //Todo remove tasks
-    res.json({ message: "section deleted" });
-  } catch (e) {
-    res.status(404).send({ error: "section not found" });
+  if (!mongoose.Types.ObjectId.isValid(sectionId)) {
+    res.status(400);
+    throw Error("Id format not correct");
   }
+
+  const section = await Section.findByIdAndDelete(sectionId);
+
+  if (!section) {
+    res.status(404);
+    throw Error("section not found");
+  }
+
+  res.json({ message: "section deleted" });
 });
 
 export {
